Moved uuid related utilities around. Added NameHistoryFetcher, though unused currently.
Improved player stats import:
if we find a playername-uuid pair with an uuid for which we already have
statistics imported, we now merge the already imported stats with the
stats for the other found player name.
This can occur if a player, for which we collected an uuid via
UUIDCollector has already changed his name and then joined the server
and created a new stats entry.