Fixed: Team names are no longer getting loaded correctly from the language file.
Since we load all classes before loading the language file now, the team names setup during the initialization of the Lobby enum are not yet available at this point.
Instead of loading the team names during the initialization of the enum, we load them when they are requested.

Also: Minor formatting.