Fixed: Since MC 1.16 the movement speed instead the lobby and game is increased.
Since MC 1.16 the walk speed is determined by both the walk speed value and the movement speed attribute.
Our code for resetting the attributes to their default values is flawed (there is currently no API to retrieve the player-specific default attribute values). This results in the players having an increased walking speed.

To resolve this for now, we always explicitly reset the walk speed now. This will then also automatically update the movement speed attribute accordingly.