Fixed: Egg hatching for grenades was no longer getting cancelled since MC 1.16.
The issue is that Spigot/Minecraft changed the order of the
ProjectileHitEvent (which removes the grenade gadget) and other events,
such as the PlayerEggThrowEvent which is used to disable hatching for
grenade gadgets. Since the grenade gets removed before the
PlayerEggThrowEvent is called now, we could no longer check if a given
egg entity is a grenade.

We now delay the grenade removal for one tick.