Fixed: The final airstrike marker was not getting removed during plugin disable.
We now remove the final airstrike marker as part of the airstrike
disposal (similar to how it is already the case for the orbital strike).

Also made some minor improvements: We were previously using
ConcurrentHashMaps to keep track of task ids, which has been replace
with regular HashMaps which should be sufficient.

We also no longer need to separately keep track of a player's final
airstrike and orbital markers. These are tracked as part of the
airstrike and orbital strike gadgets.